requirements_test.txt
-r requirements.txt
pytest>=2.6.4,<3.0.0
pytest-django>=2.8.0,<3.0.0
